Fix python 3.8 tests by resolving duplicate inheritence ambiguity
Python can't resolve the order of methods when a class extends the same base.

In this case we have `class Array(Generic[t])` and then the Floats/Ints types which extend Array and `Generic[t,...]`

Looking at the Array type it appears the argument it wants is the number of dimensions, so fill Floats/Ints types with Literal[ndim] in the implementation.

Specifying the type for our Array usage disambiguates the Generic types allowing python to derive the correct call order.