better return type for TensorflowWrapper
I went back and forth on this a bit, whether to use Array or untyped model, or what, and have the TypeVar's on the return type is useful because it throws an error if you don't specify a type for the variable it's assigned to.

Because the generic type is resolved by the annotation, you get more specifc types or an error. Win!