Concatenate parameters and only call optimizer once
PyTorch and other libraries apply gradient clipping over all the params,
where we were clipping each single parameter. This makes our hyper-parameters
hard to compare, and it also makes our Adam optimizer inefficient.

Our solution here is to concatenate the parameters prior to optimization,
and then split them back apart. We need to apply this to the MXNet and
Tensorflow shims too. The implementation could be a bit neater; we should
probably have it in a utility, or possibly we can make the optimizer itself
do the work?