Optimize parameters in bulk (#348)
* Avoid divide-by-zero in cuda reduce_mean

* Avoid divide-by-zero in some Ops kernels

* Concatenate parameters and only call optimizer once

PyTorch and other libraries apply gradient clipping over all the params,
where we were clipping each single parameter. This makes our hyper-parameters
hard to compare, and it also makes our Adam optimizer inefficient.

Our solution here is to concatenate the parameters prior to optimization,
and then split them back apart. We need to apply this to the MXNet and
Tensorflow shims too. The implementation could be a bit neater; we should
probably have it in a utility, or possibly we can make the optimizer itself
do the work?

* Fix finish_update

* Bulk optimize tensorflow and mxnet (untested)

* Fix pytorch shim finish_update

* Fix finish_update

* Update test

* Fix finish_update in shims

* Fix type errors

* Format

* Fix finish_update

* Fix test tensorflow wrapper

* Distinguish device_type and device_id when passing to shims

* Fix shim.to_device