with_nvtx_range: pass through shim of the wrapped layer
If the wrapped layer has a shim, an outer layer may need access
to it. For example, TransformerModel from spacy-transformers
needs access to the the shim of HWRapper to get the piece
tokenizer.

This change sets the shim property of the wrapper to be the shim
of the wrapped layer.