Add context manager and layer wrapper for marking NVTX ranges (#526)
* Add util.use_nvtx_range context manager

This context manager marks a scope as an NVTX range, which can be
used in CUDA profiling.

* Add with_nvtx_range layer

Layer that wraps any layer and marks the `forward` and `backprop` passes
as an NVTX range. This can be helpful when profiling GPU performance of
a layer.

* with_nvtx_range: pass through shim of the wrapped layer

If the wrapped layer has a shim, an outer layer may need access
to it. For example, TransformerModel from spacy-transformers
needs access to the the shim of HWRapper to get the piece
tokenizer.

This change sets the shim property of the wrapper to be the shim
of the wrapped layer.

* NVTX ranges: add fixes by Sofie

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>