Wrap layer callbacks in with_{debug,nvtx_range}
Before this change, with_{debug,nvtx_range} instrumented a `Model`'s
init/forward/backprop by wrapping it in another `Model`. However, this
turns out to be tricky, since in some cases we have to propagate
information from the wrapped nodes (e.g. references) while in other
cases we shouldn't.

With this change, we instead add instumentation by wrapping a `Model`'s
init and forward callbacks. This makes the nodes in the graph appear as
they were before wrapping, but adds behavior in the wrapped callbacks.

The with_{debug,nvtx_range} still returns `Model` to maintain
compatibility with the existing API. Though the returned `Model` is the
`Model` that is wrapped.