mixed-precision: do not backprop Inf/NaN out of PyTorch layers
With mixed-precision enabled in a PyTorchWrapper, an update is skipped
and the gradient scale is adjusted when one or more of the gradients is
Inf/NaN. However, Inf/NaN gradients were still backpropagated into
preceding Thinc layers.

This change fixes this issue by clearing gradients of inputs to
PyTorchWrapper if Inf/NaN occurs in any of the gradients.