chain: also pass through Y when layer without nO in initialization
chain only passed through Y when a layer has nO set to None (nO is not
yet inferred). However, some layers do not have a nO dimension (e.g.
tuplify), but may have inner layers that still need Y to be properly
initialized.

This change changes chain initialization, so that it passes through Y
when: 1. nO is not inferred *or* nO is not a dimension of the layer.