Add assert that size hasn't changed for reduce mean backprop (#572)
* Add assert that size hasn't changed for reduce mean backprop

Behavior of this with numpyops seems wrong, as instead of giving an
error it produces nans, as though it's accessing out of bounds memory or
something. See https://github.com/explosion/spaCy/pull/9669.

* Add clean checks for all reduce functions

This also saves the size in a variable so the actual data can be garbage
collected.

* Add minimal test

There are no tests for most of the reduce layers...

* Add size consistency check to softmax

I think the math won't work out anyway if this is inconsistent, but
checking here makes the cause of errors more explicit.

* Use a decorator for consistent backprop

This is cleaner than copying the other code. Maybe this decorator could
go on the top of a layer too to be even cleaner?

* Make backprop decorator work on forward

This simplifies things a bit at the point of usage.

* Add tests for all reduce methods

* Update thinc/util.py

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>

* New approach to consistency checking

Currently only in reduce_mean while I get feedback

* Restructure arrayinfo creation

* Formatting

* Replace decorator with ArrayInfo everywhere

* Update thinc/util.py

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

* Rename ainfo var

* Relax typing

* Change dtype to not pick up name

* Remove leftover reference

Co-authored-by: Adriane Boyd <adrianeboyd@gmail.com>
Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>