_custom_kernels: make all arguments kwarg-only
Some kernels take multiple arguments, which are all arrays with the same
shape. It's easy to accidentally mix up the order and introduce bugs in
that way. This change makes all arguments kwarg-only, so that we always
specify what we intend to pass at the call site.

Since _custom_kernels is an internal module, this does not result in any
public API changes.