GradScaler: raise ValueError when scaling non-CUDA tensors
Before this change, using gradient scaling on a non-CUDA tensor would
trigger an assertion. However, it is possible to trigger this error
outside Thinc by enabling mixed-precision training on a CPU. So this
should be a proper exception rather than an AssertionError.

Besides raising a ValueError, the error message is also extended to
describe how the error can be avoided.