Lazily fail on PyTorch versions without gradient scaling
We checked that gradient scaling is supported when construction a
PyTorchGradScaler. However, this gives issues when someone uses a
model that was trained with gradient scaling. In such cases, it's
safe to contruct a grad scaler, since it is not used. This change
moves the check to the actual scaling.

Also remove the check that verifies that mixed-precision scaling is
available (when enabled) from the constructor of PyTorchShim. PyTorch
will at most give a warning when trying to autocast when there is
no support.