Mixed-precision training: improve checks (#624)
* GradScaler: raise ValueError when scaling non-CUDA tensors

Before this change, using gradient scaling on a non-CUDA tensor would
trigger an assertion. However, it is possible to trigger this error
outside Thinc by enabling mixed-precision training on a CPU. So this
should be a proper exception rather than an AssertionError.

Besides raising a ValueError, the error message is also extended to
describe how the error can be avoided.

* Lazily fail on PyTorch versions without gradient scaling

We checked that gradient scaling is supported when construction a
PyTorchGradScaler. However, this gives issues when someone uses a
model that was trained with gradient scaling. In such cases, it's
safe to contruct a grad scaler, since it is not used. This change
moves the check to the actual scaling.

Also remove the check that verifies that mixed-precision scaling is
available (when enabled) from the constructor of PyTorchShim. PyTorch
will at most give a warning when trying to autocast when there is
no support.

* Add missing import