Introduce `SparseLinear_v2` to fix indexing issues
`SparseLinear` does not correctly index the gradient/weight matrix
(#752). This change fixes the indexing, so that the full matrix is
used.

To retain compatibility with existing models that use `SparseLinear`, which
works relatively well if there are not too many hash collisions, the fixed
version is renamed to `SparseLinear_v2`.

Thanks to @sriram7797 for reporting this issue!