Give schedulers access to the key, step, and last eval score
Before this change schedules were generators that generate a value for
each training step. This, however has the limitation that scheduler
cannot use other information that is available in the optimizer such
as the parameter key. This information is useful for e.g. discriminative
learning rates, where certain parameters are on a different schedule
than others.

To accommodate passing additional information, this change converts
schedules to callables. These callables are passed the training step,
the parameter key, and the last evaluation score (when available).

Traditional scalar and generated schedules are converted to callables
by the optimizer for compatibility.