Add TorchScriptWrapper_v1 (#802)
* Add TorchScriptWrapper_v1

PyTorch `Module`s can be converted to TorchScript. TorchScript has
the advantage that the model is serialized with the model parameters
in a portable manner. Deserialization (in contrast to pickling) does
not require certain Python types to be available. In fact, a TorchScript
module can be loaded in C++ without a Python interpreter.

For Thinc and spaCy, supporting TorchScript has two main benefits:

1. Since the model is serialized, we don't run into the issue that
   we have with the PyTorch wrapper where we have to construct the
   PyTorch model before deserializing its parameters but we can only
   know the model shapes through deserialization.

2. When a model is rewritten for e.g. quantization, it is unwieldy to
   construct the rewritten model by hand. So, with the TorchScript
   wrapper we would first have to construct the original model and
   then reapply the graph transformation. This is quite bad for
   transformations that reduce the model size, since we end up
   temporarily allocating all parameters. This is not an issue
   with TorchScript, since we serialize the rewritten model.

* Add fixes suggested by @svlandeg