`with_flatten`: make the layer input/output types symmetric
The idea of the `with_flatten` layer is that it flattens a nested sequence, passes
it to the wrapped layer, and then unflattens the output of the wrapped layer.

However, the layer was asymmetric in that it passes a list to the wrapped layer,
but expects back an XP array. This breaks composition with other Thinc layers,
such as `with_array`.

This change makes `with_flatten` symmetric, in that the inputs/outputs of the
`with_flatten` and the wrapped layer are symmetric.