Move from callables to a `Schedule` class
The new learning rate functionality used `Callable`s. However, the issue
with callables it that they cannot be pickled. This is problematic,
because schedules can end up in spaCy pipelines (e.g. through the
optimizer associated with the `Language` object).

This change solves this issue by refactoring the schedules into regular
objects. This now works similar to Thinc `Model`s -- there is a new
`Scheduler` class which can be constructed with composition.

I tested the changes with spaCy and pickling as well as usin existing
configurations works.