Give schedules access to the key, step, and last eval score (#804)
* Give schedulers access to the key, step, and last eval score

Before this change schedules were generators that generate a value for
each training step. This, however has the limitation that scheduler
cannot use other information that is available in the optimizer such
as the parameter key. This information is useful for e.g. discriminative
learning rates, where certain parameters are on a different schedule
than others.

To accommodate passing additional information, this change converts
schedules to callables. These callables are passed the training step,
the parameter key, and the last evaluation score (when available).

Traditional scalar and generated schedules are converted to callables
by the optimizer for compatibility.

* Fix use of the `t` parameter where used in the schedules

Also add tests, so that doesn't break again.

* Fixes from @shadeMe

* Call _schedule_args once

* Make Optimizer.step private

* Fix two missed step uses in tests

* Float fix

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>

* Fix schedule call

* Move `ScheduleCallable` to `thinc.types`

* Move from callables to a `Schedule` class

The new learning rate functionality used `Callable`s. However, the issue
with callables it that they cannot be pickled. This is problematic,
because schedules can end up in spaCy pipelines (e.g. through the
optimizer associated with the `Language` object).

This change solves this issue by refactoring the schedules into regular
objects. This now works similar to Thinc `Model`s -- there is a new
`Scheduler` class which can be constructed with composition.

I tested the changes with spaCy and pickling as well as usin existing
configurations works.

* Remove stray `runtime_checkable` import

* Apply suggestions from code review

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>
Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>