Avoid h2d - d2h roundtrip when using `unflatten` (#861)
* Avoid h2d - d2h roundtrip when using `unflatten`

`unflatten` converts its `lengths` argument to a NumPy array, because
CuPy's `split` function requires lengths to be in CPU memory. However,
in various places in Thinc, we copy the lengths array to GPU memory when
CupyOps is used. This results in an unnecessary roundtrip of the lengths
array (host to device -> device to host). One of these roundtrips
(array `list2array`) showed up in profiles of the biaffine parser.

This change fixes some length array allocations to avoid the round trip.

* Add a comment to `with_ragged` to avoid confusion about memory allocation