Implement `pad` as a CUDA kernel (#860)
* Implement `pad` as a CUDA kernel

`Ops.pad` was a fairly slow operation on GPU. It iterates over all
sequences and copies each sequence into the padded array. This results
in a lot of kernel launches. In the biaffine parser, padding the inputs
was more costly than applying the biaffine layers.

This change optimizes the `pad` op using a custom CUDA kernel. The
kernel get an array of pointers to the CuPy arrays that are provided as
a list. The output array is then filled, parallelizing over the 'time
steps'. This should provides the largest amount of parallelism, since
we usually have n_steps * hidden_size to parallelize over.

* Rename variables for clarification

* Better validation of incorrect rounding

* Simplify rounding using modular arithmetic, add test